+++ Changelog.md
@@ -1,3 +1,12 @@
+### 3.2.0 Development
+[Full Changelog](http://github.com/rspec/rspec-expectations/compare/v3.1.0...master)
+
+Enhancements:
+
+* Add `block_arg` method to custom matcher API, which allows you to
+  access the block passed to a custom matcher, if there is one.
+  (Mike Dalton, #645)
+
 ### 3.1.0 / 2014-09-04
 [Full Changelog](http://github.com/rspec/rspec-expectations/compare/v3.0.4...v3.1.0)